refactor: #3
