+++ bot.js
@@ -4,7 +4,7 @@ const { Telegraf, Stage, session, BaseScene } = require('telegraf');
 const I18n = require('telegraf-i18n');
 const path = require('path');
 const { DefaultActionController } = require('./src/controllers/default-actions.controller');
-const { RegisterController } = require('./src/controllers/choice-nickname.controller');
+const { RegisterController } = require('./src/controllers/register.controller');
 const { MenusController } = require('./src/controllers/menus.controller');
 const { UserSettingController } = require('./src/controllers/user-setting.controler');
 const { UserPostController } = require('./src/controllers/user-post.controller');
