+++ src/controllers/register.controller.js
@@ -1,3 +1,4 @@
+const { RegisterView } = require('../view/register-view');
 const { View } = require('../view/view');
 
 class RegisterController {
@@ -5,27 +6,26 @@ class RegisterController {
 	#postServices;
 	#commentServices;
 	#view;
+	#registerView;
 	constructor(UserServices, PostServices, CommentServices) {
 		this.#userServices = UserServices;
 		this.#postServices = PostServices;
 		this.#commentServices = CommentServices;
 		this.#view = new View(UserServices, PostServices, CommentServices);
+		this.#registerView = new RegisterView(UserServices, PostServices, CommentServices);
 	}
 
 	async firstChoiceNickname(ctx) {
-		console.log(ctx.message.text);
 		const nickname = await this.#userServices.getUserByNickName(ctx.message.text);
-		console.log('nickname: ', nickname);
 		if (nickname) {
-			this.#view.busyNickname(ctx);
+			this.#registerView.busyNickname(ctx);
 		} else {
 			const newUser = await this.#userServices.createUser({
 				name: ctx.from.first_name,
 				nickName: ctx.message.text,
 				telegramId: ctx.from.id,
 			});
-			console.log('newUser: ', newUser);
-			this.#view.firstChoiceAgeView(ctx);
+			this.#registerView.firstChoiceAgeView(ctx);
 		}
 	}
 
@@ -33,12 +33,12 @@ class RegisterController {
 		await this.#userServices.updateUser(ctx.from.id, {
 			age: parseInt(ctx.message.text),
 		});
-		this.#view.firstChoiceSexView(ctx);
+		this.#registerView.firstChoiceSexView(ctx);
 	}
 
 	async firstChoiceSex(ctx) {
 		if (ctx.message.text !== (ctx.i18n.t('buttons.men') || ctx.i18n.t('buttons.women'))) {
-			this.#view.firstChoiceSexView(ctx);
+			this.#registerView.firstChoiceSexView(ctx);
 		} else {
 			await this.#userServices.updateUser(ctx.from.id, {
 				sex: ctx.message.text,
