+++ src/view/view.js
@@ -18,26 +18,6 @@ class View {
 		this.#keyboards = new Keyboards();
 	}
 
-	async firstChoiceNicknameView(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.firstChoiceNickname'));
-		ctx.scene.enter('firstChoiceNickname');
-	}
-
-	async busyNickname(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.busyNickname'));
-		ctx.scene.enter('firstChoiceNickname');
-	}
-
-	async firstChoiceAgeView(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.firstChoiceAge'));
-		ctx.scene.enter('firstChoiceAge');
-	}
-
-	async firstChoiceSexView(ctx) {
-		ctx.reply(ctx.i18n.t('phrases.firstChoiceSex'), this.#keyboards.choiceSex(ctx));
-		ctx.scene.enter('firstChoiceSex');
-	}
-
 	async mainMenu(ctx) {
 		const user = await this.#userServices.getUserByTgId(ctx.from.id);
 		ctx.session.user = user;
